Do not make MaliciousSiteBlockerWebViewIntegration public
