--- browser-api/src/main/java/com/duckduckgo/browser/api/MaliciousSiteBlockerWebViewIntegration.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2024 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.browser.api
-
-import android.net.Uri
-import android.webkit.WebResourceRequest
-import android.webkit.WebResourceResponse
-
-interface MaliciousSiteBlockerWebViewIntegration {
-
-    suspend fun shouldIntercept(
-        request: WebResourceRequest,
-        documentUri: Uri?,
-        onSiteBlockedAsync: () -> Unit,
-    ): WebResourceResponse?
-
-    suspend fun shouldOverrideUrlLoading(
-        url: Uri,
-        webViewUrl: Uri?,
-        isForMainFrame: Boolean,
-        onSiteBlockedAsync: () -> Unit,
-    ): Boolean
-
-    fun onPageLoadStarted()
-}
